+++ src/Endpoint/AccountMeTwoFactorAuthEndpoint.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace DigitalCz\DigiSign\Endpoint;
 
 use DigitalCz\DigiSign\Resource\BaseResource;
+use DigitalCz\DigiSign\Resource\TwoFactorAuth;
 
 /**
  * @extends ResourceEndpoint<BaseResource>
@@ -14,6 +15,11 @@ final class AccountMeTwoFactorAuthEndpoint extends ResourceEndpoint
     public function __construct(AccountMeEndpoint $parent)
     {
         parent::__construct($parent, '/2fa');
+    }
+
+    public function get(): TwoFactorAuth
+    {
+        return $this->createResource($this->getRequest(), TwoFactorAuth::class);
     }
 
     /**